refactor: rename RenderIcon to Pascal case
